# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api Documentation for Website",
        "title": "Website Api",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8085",
    "basePath": "/",
    "paths": {
        "/admin/orders": {
            "get": {
                "description": "Mengambil semua order (khusus admin)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "4. Order"
                ],
                "summary": "List Orders (Admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter status order",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil daftar order",
                        "schema": {
                            "$ref": "#/definitions/order.OrderResponse"
                        }
                    }
                }
            }
        },
...
        "/order/{id}": {
            "get": {
                "description": "Mengambil detail order beserta item (hanya pemilik order atau admin)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "4. Order"
                ],
                "summary": "Get Order Detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail order",
                        "schema": {
                            "$ref": "#/definitions/order.OrderResponse"
                        }
                    }
                }
            }
        },
        "/order/{id}/payment-link": {
            "post": {
                "description": "Membuat payment link Midtrans untuk order yang masih PENDING sehingga bisa dibagikan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "4. Order"
                ],
                "summary": "Create Payment Link",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment link berhasil dibuat",
                        "schema": {
                            "$ref": "#/definitions/order.PaymentLinkResponse"
                        }
                    }
                }
            }
        },
        "/product/{id}": {
            "get": {
                "description": "Mengambil detail produk beserta variannya",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Get Product Detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail produk",
                        "schema": {
                            "$ref": "#/definitions/product.ProductDetailResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "order.OrderItemResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "product_variant_id": {
                    "type": "string"
                },
                "qty": {
                    "type": "integer"
                },
                "subtotal": {
                    "type": "number"
                }
            }
        },
        "order.OrderResponse": {
            "type": "object",
            "properties": {
                "address_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "expired_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/order.OrderItemResponse"
                    }
                },
                "payment_method": {
                    "type": "string"
                },
                "payment_url": {
                    "type": "string"
                },
                "shipping_fee": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                }
            }
        },
        "order.PaymentLinkResponse": {
            "type": "object",
            "properties": {
                "order_id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "product.ProductDetailResponse": {
            "type": "object",
            "properties": {
                "base_price": {
                    "type": "number"
                },
                "brand_id": {
                    "type": "string"
                },
                "brand_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_in_stock": {
                    "type": "integer"
                },
                "max_price": {
                    "type": "number"
                },
                "min_price": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "slug": {
                    "type": "string"
                },
                "thumbnail": {
                    "type": "string"
                },
                "total_review": {
                    "type": "integer"
                },
                "variants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/product.VariantResponse"
                    }
                }
            }
        },
        "product.VariantResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "sku": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                },
                "variant_name": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        }
    },
...
}